requirements(1).txt
azure-ai-documentintelligence==1.0.0b4
azure-storage-blob
python-dotenv
streamlit